fix: resolve 6 failing CI tests for string concat and class type inference
Fixes two root causes of CI test failures:

1. Class Type Inference (test_46, test_50):
   - When HM inference produces UnificationVar for class constructors,
     the type mapper was falling back to HashSet<DepylerValue>
   - Added detection for class constructor calls with unresolved types
   - Now generates correct type annotation: `let calc: Calculator = ...`

2. String Concatenation Cast (test_03, test_36, test_38, test_138):
   - Extended expr_is_string_type to handle Binary Add expressions
   - Added target_is_string check to prevent invalid `as i32` casts
   - String concatenation chains now handled correctly

All 6 previously failing tests now pass:
- test_03_literals_strings
- test_36_string_methods
- test_38_string_formatting
- test_46_instance_method
- test_50_method_chaining
- test_138_string_concatenation

+++ examples/custom_attributes_demo.rs
@@ -3142,11 +3142,11 @@ pub fn compute_hash(text: &str) -> Result<i32, Box<dyn std::error::Error>> {
     let mut hash_val: i32 = Default::default();
     hash_val = 0;
     for char in text.chars() {
-        hash_val = (((hash_val).py_mul(31i32)).py_add(char as u32 as i32)).py_mod(
+        hash_val = ((((hash_val).py_mul(31i32) as i32).py_add(char as u32 as i32) as i32).py_mod(
             ({ 2 } as i32)
                 .checked_pow({ 32 } as u32)
                 .expect("Power operation overflowed"),
-        );
+        )) as i32;
     }
     Ok(hash_val)
 }
